+++ src/canisters/frontend/ic-asset/src/asset/content.rs
@@ -1,4 +1,5 @@
-use crate::content_encoder::ContentEncoder;
+use crate::asset::content_encoder::ContentEncoder;
+
 use flate2::write::GzEncoder;
 use flate2::Compression;
 use mime::Mime;
