+++ src/canisters/frontend/ic-asset/src/batch_upload/semaphores.rs
