+++ src/canisters/frontend/ic-asset/src/canister_api/methods/list.rs
@@ -1,5 +1,5 @@
-use crate::asset_canister::method_names::LIST;
-use crate::asset_canister::protocol::{AssetDetails, ListAssetsRequest};
+use crate::canister_api::methods::method_names::LIST;
+use crate::canister_api::types::{asset::AssetDetails, list::ListAssetsRequest};
 use ic_utils::call::SyncCall;
 use ic_utils::Canister;
