+++ src/canisters/frontend/ic-asset/src/canister_api/methods/method_names.rs
