+++ src/canisters/frontend/ic-asset/src/canister_api/methods/mod.rs
