+++ src/canisters/frontend/ic-asset/src/canister_api/types/batch_upload.rs
@@ -1,4 +1,4 @@
-use crate::asset_config::HeadersConfig;
+use crate::asset::config::HeadersConfig;
 use candid::{CandidType, Nat};
 use serde::Deserialize;
 
@@ -32,32 +32,6 @@ pub struct CreateChunkRequest<'a> {
 pub struct CreateChunkResponse {
     /// The ID of the created chunk.
     pub chunk_id: Nat,
-}
-
-/// Return a list of all assets in the canister.
-#[derive(CandidType, Debug)]
-pub struct ListAssetsRequest {}
-
-/// Information about a content encoding stored for an asset.
-#[derive(CandidType, Debug, Deserialize)]
-pub struct AssetEncodingDetails {
-    /// A content encoding, such as "gzip".
-    pub content_encoding: String,
-
-    /// By convention, the sha256 of the entire asset encoding.  This is calculated
-    /// by the asset uploader.  It is not generated or validated by the canister.
-    pub sha256: Option<Vec<u8>>,
-}
-
-/// Information about an asset stored in the canister.
-#[derive(CandidType, Debug, Deserialize)]
-pub struct AssetDetails {
-    /// The key identifies the asset.
-    pub key: String,
-    /// A list of the encodings stored for the asset.
-    pub encodings: Vec<AssetEncodingDetails>,
-    /// The MIME type of the asset.
-    pub content_type: String,
 }
 
 /// Create a new asset.  Has no effect if the asset already exists and the content type matches.
